+++ src/Integration/UserIntegrations/GitLabIntegration.php
@@ -1903,4 +1903,8 @@ class GitLabIntegration implements PersonalizedSkillInterface
             'integration_id' => $config?->getId() ?? 'XXX',
         ]);
     }
+    public function isExperimental(): bool
+    {
+        return false;
+    }
 }