+++ src/Integration/UserIntegrations/JiraIntegration.php
@@ -651,4 +651,8 @@ class JiraIntegration implements PersonalizedSkillInterface
             'integration_id' => $config?->getId() ?? 'XXX',
         ]);
     }
+    public function isExperimental(): bool
+    {
+        return false;
+    }
 }